requirements.txt
alabaster==0.7.13
annotated-types==0.6.0
anyio==4.0.0
astroid==3.0.1
autopep8==2.0.4
Babel==2.13.1
beautifulsoup4==4.12.2
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.7
docutils==0.20.1
et-xmlfile==1.1.0
flake8==6.1.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
imagesize==1.4.1
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
platformdirs==4.0.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.16.1
pylint==3.0.2
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.23
tomlkit==0.12.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
